Add test per feedback
